+++ src/main/java/org/ayty/hatcher/api/v1/competence/jpa/CompetenceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompetenceRepository extends JpaRepository<Competence, Integer> {
+public interface CompetenceRepository extends JpaRepository<Competence, Long> {
 
 }
